+++ packages/plugins/EVM/src/state/IdentityService.ts
@@ -19,8 +19,8 @@ import { MaskX_BaseAPI } from '@masknet/web3-providers/types'
 import { ENS_Resolver } from './NameService/ENS.js'
 import { SpaceID_Resolver } from './NameService/SpaceID.js'
 import { Web3StateSettings } from '../settings/index.js'
-import { GraphQLResolver } from './NameService/GraphQL.js'
-import { ThegraphResolver } from './NameService/Thegraph.js'
+import { R2D2Resolver } from './NameService/R2D2.js'
+import { TheGraphResolver } from './NameService/Thegraph.js'
 
 const ENS_RE = /[^\s()[\]]{1,256}\.(eth|kred|xyz|luxe)\b/gi
 const SID_RE = /[^\t\n\v()[\]]{1,256}\.bnb\b/gi
@@ -177,7 +177,7 @@ export class IdentityService extends IdentityServiceState<ChainId> {
         const allSettled = await Promise.allSettled(
             names.map(async (name) => {
                 const address = await attemptUntil(
-                    [new ENS_Resolver(), new GraphQLResolver(), new ThegraphResolver()].map((resolver) => {
+                    [new ENS_Resolver(), new R2D2Resolver(), new TheGraphResolver()].map((resolver) => {
                         return async () => resolver.lookup(name)
                     }),
                     undefined,
@@ -268,7 +268,7 @@ export class IdentityService extends IdentityServiceState<ChainId> {
             results.map(async (y) => {
                 try {
                     const name = await attemptUntil(
-                        [new ENS_Resolver(), new GraphQLResolver(), new ThegraphResolver()].map((resolver) => {
+                        [new ENS_Resolver(), new R2D2Resolver(), new TheGraphResolver()].map((resolver) => {
                             return async () => resolver.reverse(y.web3_addr)
                         }),
                         undefined,
