+++ packages/plugins/EVM/src/state/NameService/Thegraph.ts
@@ -1,16 +1,16 @@
 import { NameServiceID } from '@masknet/shared-base'
-import { Thegraph } from '@masknet/web3-providers'
+import { TheGraphDomain } from '@masknet/web3-providers'
 import type { NameServiceResolver } from '@masknet/web3-shared-base'
 import { ChainId } from '@masknet/web3-shared-evm'
 
-export class ThegraphResolver implements NameServiceResolver {
+export class TheGraphResolver implements NameServiceResolver {
     get id(): NameServiceID {
         return NameServiceID.ENS
     }
     async reverse(address: string) {
-        return Thegraph.reverse(ChainId.Mainnet, address)
+        return TheGraphDomain.reverse(ChainId.Mainnet, address)
     }
     async lookup(name: string) {
-        return Thegraph.lookup(ChainId.Mainnet, name)
+        return TheGraphDomain.lookup(ChainId.Mainnet, name)
     }
 }
