+++ packages/web3-providers/src/TheGraph/apis/DomainAPI.ts
@@ -1,10 +1,10 @@
-import type { ChainId } from '@masknet/web3-shared-evm'
-import type { DomainAPI } from '../entry-types.js'
-import { API_URL } from './constants.js'
-import { fetchJSON } from '../entry-helpers.js'
 import { first } from 'lodash-es'
+import type { ChainId } from '@masknet/web3-shared-evm'
+import { API_URL } from '../constants.js'
+import type { DomainAPI } from '../../entry-types.js'
+import { fetchJSON } from '../../entry-helpers.js'
 
-export class Thegraph_Domain_API implements DomainAPI.Provider<ChainId> {
+export class TheGraphDomainAPI implements DomainAPI.Provider<ChainId> {
     async lookup(chainId: ChainId, name: string): Promise<string | undefined> {
         const response = await fetchJSON<{
             data: {
