+++ packages/web3-providers/src/TheGraph/index.ts
@@ -1,0 +1,1 @@
+export * from './apis/DomainAPI.js'
